Fix clippy warnings: useless format!, unnecessary cast, missing fields
- Replace format!("fcvm-user") with "fcvm-user".to_string()
- Remove unnecessary `as i32` cast on as_raw_fd() return
- Add missing forward_localhost and user fields to test_library_api RunArgs

+++ tests/test_forward_localhost.rs
@@ -32,7 +32,7 @@ async fn test_forward_localhost() -> Result<()> {
                 tv_usec: 0,
             };
             libc::setsockopt(
-                std::os::unix::io::AsRawFd::as_raw_fd(&listener) as i32,
+                std::os::unix::io::AsRawFd::as_raw_fd(&listener),
                 libc::SOL_SOCKET,
                 libc::SO_RCVTIMEO,
                 &tv as *const _ as *const libc::c_void,
